+++ src/test/java/roomescape/controller/ViewControllerTest.java
@@ -1,0 +1,39 @@
+package roomescape.controller;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class ViewControllerTest {
+
+    @DisplayName("/admin get 요청 시 응답할 수 있다")
+    @Test
+    void should_response_200_when_request_admin_page() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("웰컴페이지 요청 시 응답할 수 있다")
+    @Test
+    void should_response_200_when_request_welcome_page() {
+        RestAssured.given().log().all()
+                .when().get("/")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("/admin/reservation get 요청 시 응답할 수 있다")
+    @Test
+    void should_response_200_when_request_reservation_page() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
